+++ data/user/src/main/java/com/hilingual/data/user/dto/reponse/FollowerResponseDto.kt
@@ -1,0 +1,10 @@
+package com.hilingual.data.user.dto.reponse
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FollowerResponseDto(
+    @SerialName("followerList")
+    val userList: List<UserItemDto>
+)
