+++ data/user/src/main/java/com/hilingual/data/user/dto/reponse/FollowingResponseDto.kt
@@ -1,0 +1,10 @@
+package com.hilingual.data.user.dto.reponse
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FollowingResponseDto(
+    @SerialName("followingList")
+    val userList: List<UserItemDto>
+)
